Making status update handling more flexible
We previously had a page-specific bit of logic in the
startGraphStatusUpdates::updateStates function that prevented us from
reusing the status polling mechanism to update other parts of the web
UI. This change makes that bit parametrizable and moves the previous
page-specific bits to where they belonged.

This changes will allow us to properly update the "list" view of the
physical graph on the session page, which up to now only displays the
drops' OIDs but not their states.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>