Supporting PG modification functions
Adding this support is something we currently need for the SUMMIT demo.
Modifications to the PG shouldn't be necessary in principle, but given
that we are lacking supporting during the translation and partitioning
phase to apply any further logic, this is the easiest and most generic
non-intrusive way I couldn't think of to allow any further PG
customization.

In non-HPC contexts, this is equivalent to adding an additional command
to the "dlg unroll | dlg partition | dlg map | dlg submit" pipe we
usually use to submit LGs into a running system.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>