Using SELF communicator
It turned out that at the end of the day we needed to use a SELF
communicator for our experiments, so let's change it to be so. This code
is highly specific to our SUMMIT demo experiments, so at some point we
should remove it and/or abstract it away.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>