Automatic transition to COMPLETED/ERROR
Generic AppDROPs (i.e., those not extending BarrierAppDrop) currently
have too much responsibility, as they need to manually set their
execution state, set their base drop state, and notify that their
execution has finished via _notifyAppIsFinished. This was clearly a poor
design, and this commit is trying to alleviate that a bit, automatically
setting the base drop status of the application drop based on the
execution status set by the user.

This was noted when exercising the built-in CRCStreamApp, which lacked
a manual base drop state transition.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>