Fixing remote linking to streaming consumers
When adding streaming consumers to a data drop we were checking whether
the drop given as a parameter "looked" like a streaming consumer by
checking for the presence of some methods. However, when executed in the
context of a multi-nm graph deployment, the object given to this
function can be a drop proxy, therefore triggering a remote call to the
remote actual drop, which might or might not exist at that point in time
(drops are created independently in parallel in all NMs).

This particular scenario was never experienced before, mainly because
streaming consumers/inputs are so rare in our exercises, but also
because there was no unit test exercising this particular functionality.
Now there is.

The fact that we are removing these checks shouldn't be too worrying: we
don't have these kind of checks for other linkage operations
(addProducer, addConsumer, etc) and we have survived without them.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>