start_dfms_cluster using new Remote classes
Now that we have the new Remote classes available, nicely encapsulating
the different technologies used to remotely coordinate the startup of
the daliuge cluster, we are porting the main function of the
start_dfms_cluster module to use these new objects.

Along with this porting also come a few small simplifications to the
structure of the code, specially the startup of the Node Managers, which
is identical in both the single- and multiple-island cases. More
improvements can be made over the logic of the code to further improve
the maintainability of the code, but overall things are looking much
better now.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>