New remote classes and module
The aim of the new Remote classes (which are contained in the new
remotes module) is to abstract away all the necessary coordination
between the different processes that make up the daliuge cluster started
by start_dfms_cluster. So far we have intermingled MPI-based
coordination code in between the business logic of start_dfms_cluster,
making it difficult to track, and specially to replace. Abstracting that
logic into these Remote classes makes it easier to follow what is going
on in start_dfms_cluster.

Two implementations are offered: MPIRemote, which implemented the
current logic found in start_dfms_cluster, and SlurmRemote, which works
based on SLURM-provided variables. Common functionality (which is most
of it) is provided by a common base class.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>