Making cmdline to algo parameters parsing reusable
This will be needed if we don't want to repeat ourselves in the
start_dfms_cluster script, where we also want to accept algorithm
parameters.

While doing this I realized the list of algorithm parameter names was
out of date and didn't match the actual expected names. I fixed this for
now, but names, type information and default values should obviously be
unified in the future to avoid further mistakes.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>