New dlg.deploy.common module
This module is meant to contain deployment logic that is common to a
number of tools. For the time being we have two main functions: submit,
to submit a physical graph to a manager, and monitor_sessions to block
until one or all sessions in a manager have finished.

A new unit test module makes sure this functionality works as intended.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>