Allowing LG construction from python list
The only way to construct an LG object previously was by reading JSON
data from a file, which was either already opened or was a string
indicating a file to be opened for reading. In a few cases we already
have the logical graph data in memory though, so there's no need to
perform this JSON loading step.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>