pg_generator.partition does the whole partitioning
The previous implementation of this method left the caller with the
responsibility to actually get the final form of the partitioned
physical graph, which is done through an awkward call to the to_pg_spec
method of the resulting object. This calls were repeated in a number of
locations, and were easy to miscode, so it's better to bring them
together under the common umbrella of pg_generator.partition, which now
returns the final list of drops that make out a physical graph.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>