Requesting correct number of partitions
Previously we were counting the number of DIMs as partitions to be
requested. This is not what we need though, as the number of partitions
corresponds to the NMs only.

This was actually causing problems when running a partitioning with
mysarkar, as it internally calculates its own number of partitions, and
only reduces the number of partitions through merging if the original
count is bigger than the required one.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>