Improving the shutdown protocol for managers
Shutting down a manager *process* was previously possible only through
signal handling, which correctly shut down the manager itself, and the
enclosing rest server that exposed it to the outside world. Now the
shutdown logic is managed entirely by the rest server, who shuts down
the manager and itself, and exposes this functionality via a REST API
call, giving the possibility to send requests for shutdown remotely. The
corresponding functionality has also been added to the REST client.

Yes, not ideal from the point of view of security, but we haven't put
attention to security at any point in time and this is no exception.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>