Moving from --loc to --interface
The old --loc option required some knowledge to be hardcoded into the
start_dfms_cluster, which was not very nice, and also didn't allow for
selecting specific interfaces via netifaces. The new --interface option
is much versatile, as it allows specifying an index instead, works with
both mechanisms, and doesn't require hardcoding data in the script. To
accommodate for the new option the ifconfig-based mechanism had to be
re-written of course.

Also added some security on top of the --check-interfaces option so it
always tries to report values using both methods, even if one fails.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>