Log exceptions explicitly before giving up
If the drop manager fails to be created (e.g., due to startup problems)
we want to be able to log more explicitly why this happened in the first
place, instead of just relying on the exception being eventually printed
out into stderr.

In particular, during our SUMMIT demo, and in general when using the
start_dfms_cluster script, we might not get the stderr output
necessarily captured.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>