Continue monitoring forever
We have seen at least once a situation in which the DIM couldn't contact
all NMs when polling their statuses, which translated into an exception
being thrown in the monitoring thread. This provoked the monitoring
thread to finish, and therefore the main thread continued, shutting down
all services and exiting prematurely.

We don't want to run into those situations again, so we are simply
capturing errors but then trying again.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>