Deducing rank from global list
When trying the previous solution in SUMMIT I saw that the node mapping
was not always the same through different invocations of jsrun, so the
position of a node (and therefore its IP) in the global list of IPs
given via DALIUGE_CLUSTER_IPS cannot be assumed to be the same. A
stronger check is then to let each node deduce its rank by looking the
position of its IP (i.e., my_ip) in the global list given via
DALIUGE_CLUSTER_IPS, which is common to all nodes.

With this change deployments of 6 and 200 nodes start correctly, so I
see no reason to believe that a full deployment will not work.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>