Support for string-based volume bindings
Additional bindings have been specified until now as a list of string,
but since the logical graph editors don't support this type yet, users
are constrained to writing strings only. With this change we are
accepting volume bindings not only in list form, but also in string
form, in which case separate bindings should be comma-separated.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>