OpenAPI files for dlg managers
This new version brings only not only the common manager REST API, but
also the specific APIs exposed by the node and composite managers as
well. The content previously found in daliuge_common.yaml is now in
manager_common.yaml, but has been greatly revamped to more comments and
documentation, as well as schemas describing data structures used in
that document and in the others.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>