Adding tests for OpenAPI-related functionality
This new test first runs a validation of the files to make sure they
adhere to the standard, and then generates python bindings for each of
them. Finally, a python script uses the generated code to run a few
operations in a live instance (a NM plus a DIM) to make sure the calls
are actually making it to the server.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>