Current top-level code now under daliuge-rutime
This is the very first commit in a series of commits that attempt to
break down the single daliuge python package into three different ones:
daliuge-common, daliuge-runtime and daliuge-translator. This comes out
of the necessity to reduce the number of dependencies on the translator
code, which itself is quite lean and fairly independent of the runtime
part of daliuge. There are only a few bits of code that are really
shared across both parts of daliuge, which is what will go into
daliuge-common, which will be a dependency for both.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>