First cut at having split repositories
This first commit firstly renames the "daliuge" package to
"daliuge-runtime". It then creates a new "daliuge-common" package, and
takes a handful of functions and the tool module down from
daliuge-runtime into daliuge-common. All tests still remain together
outside the packages structure, but they have been modified to work with
the new two-package structure, as well as the rest of the code.

This change is as non-invasive as possible but without breaking the
tests. A more invasive change would include moving all the modules
currently under daliuge-runtime/dlg into daliuge-runtime/dlg/runtime,
but this is better done as a separate step.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>