Breaking down tests
The bulk of the tests are now under daliuge-runtime, but I took all of
test.dropmake and put it under daliuge-translator. This will help us
test effectively if the code under daliuge-translator works, or not,
without daliuge-runtime being there for us.

At the moment all tests pass if all three packages are installed. The
only modification I had to do was to break down the test_tool.py test in
daliuge-runtime, as it was using resources that are now under
daliuge-translator (where the individual unit test was moved to).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>