More code being moved down to daliuge-common
These are bits that are used by the translator one way or another. Since
we started moving more and more code down to daliuge-common it is
sensible to actually separate it into different modules, but still
import all names into the dlg.common package for easier import
elsewhere. We are also being careful with not breaking backwards
compatibility, so every time we mode some code down the dlg.common we
also import that name into the original module.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>