Adjust tool's commands to their new home
Most commands are actually hosted in the daliuge-translator actually,
and in order for tests to run we need to point to the correct location.

We are also taking the opportunity to adjust which commands are exposed
by the tool depending on which packages are installed.  This is a bit of
a circular dependency, but until we devise a better solution it will
have to do.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>