Adjusting Travis for more complex job combinations
Now that we have separate packages in the same repo we want to make sure
we can run daliuge-translator and its tests without having
daliuge-runtime installed, and viceversa. Until now we have had an
implicit job matrix in our Travis definition, but given the more complex
test combinations we want to do it's time to have an explicit one.

In this commit we make the matrix explicit, using environment variables
to denote what combination we are trying out on each job. Most jobs run
still "the old way", installing all packages, running all tests, and
reporting all coverage, but we also test at least once that we can run
daliuge-runtime and daliuge-translator independently.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>