Splitting dependencies for each package
This was one of the main aspects we wanted to address with the splitting
of the code: the daliuge runtime has a big list of dependencies that are
not needed by the translator, and splitting the code would allow us to
keep track of these dependencies separately, making translator-only
installations leaner and easier.

I also took the chance to remove the "hack" to install numpy. These days
I think we don't really run into issues anymore, as most times users
should be installing wheels anyway, so let's simply remove the special
treatment numpy was receiving, which wasn't a very clean operation.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>