Solving the dlg command loading issue
Using setuptool's entry points we can on the one hand register 'plugins'
that external packages implement for a particular entry point string,
which are then loaded by the base package exposing that entry point.
This is similar to Eclipse's RPC mechanism for extending plug-ins, and
works nicely to add commands to the 'dlg' command-line tool, without it
having to know what packages is going to contribute commands, or how.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>