Don't flatten names into dlg package anymore
This flattening is not really useful anymore, because in order to
achieve it one has to import dlg.runtime, which will "magically" make
dlg.delayed appear for example. This flattening was here only to try to
avoid backwards incompatibility, but that seems unavoidable. Instead,
some release notes will have to be written on how to do this in the
future.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>