Fix small problems with documentation generation
On the one hand we were still trying to auto-document dlg.delayed, which
is now under dlg.runtime; on the other we were missing (for a long time
though, and not necessarily due to the package separation) a few
top-level imports needed by some of the modules that we auto-document.
In particular the dlg.dropmake module contents did not have their
documentation automatically generated because numpy was missing from the
list of packages to mock.

This is a documentation-only change, so let's [skip-ci]

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>