Add new u2s function
This is different to common.b2s (always turns bytes into a str instance)
and six.b (always turns a str instance into bytes): this one turns a
text object (str in py3, unicode in 2.7) into a str object. One does not
usually need this, but we have now a use case.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>