Remove old experiment-related scripts and data
This set of scripts, Java code, CSV files and and jupyter notebook were
used during the initial scalability benchmarks we ran in Pawsey and
Tianhe-II back in 2016. These files have no place in our codebase
though, so they are being removed in this commit. If anyone needs them
they are in the git history.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>