Remove spead receiver drop
This application drop was initially written to demonstrate that we could
use spead2 and daliuge (dfms at the time) together. This application has
never seen usage however, its tests are always skipped, and therefore is
simply dead code.

If needed in the future, this code could be resurrected, although it's
nothing special and can easily be reproduced (probably in a better
shape).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>