Simplify member variable names for easier reading
We already are within the context of an ZMQPubSubMixIn class so there's
no need for additional "zmq" prefixes/suffixes in member names, it makes
the code a bit more difficult to follow otherwise.

The names of the different members also didn't properly describe what
they did. The new names instead reflect a bit better the intent behind
them. There was also a log statement that read "Receiving" instead of
"Publishing", as it most probably got lost in this confusion of names.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>