Create a single ZeroMQ Context per Node Manager
This commit introduces the creation of a unique ZeroMQ Context instance
at the level of the NodeManager class that can be used instead of the
per-class Context instances we have been using in the ZeroPubSubMixIn,
ZeroRPCClient and ZeroRPCServer classes. In order to be usable from all
of these, this context object must be created using zerorpc's Context
class, which contains a few additions on top of a ZeroMQ's Context, and
that are used by ZeroRPC's classes.

Instead of managing their own Context instances, now the ZMQPubSubMixIn,
the ZeroRPCClient and the ZeroRPCServer classes all share the new,
unique Context instance created by the NodeManager classes, who takes
care of creating it and terminating it.

This change not only makes the code easier to maintain (because there is
less resource management in general), it also should reduce the number
of threads used by the node manager, since each Context will create at
least one I/O thread.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>