Use terminate_or_kill to stop daemon process
We previously manually sent signals to the daemon process, but failed to
properly await for it to finish, and to send a more drastic KILL signal.
Now that we have our new ExistingProcess class we simply wrap the daemon
process into one of these and reuse our terminate_or_kill method, which
does all the proper checks and cleanups for us.

An effect of reusing terminate_or_kill is that we will use loggers, but
without having configured the codebase to use the instrumented DlgLogger
class. Thus we need to ensure that our logging configuration does *not*
include the extra record attributes included by DlgLogger.

This addresses LIU-51.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>