Skip connection monitoring for connected endpoints
In 67b58ac we introduced a mechanism through which we ensure an actual
connection is made when setting up a subscription from a SUBSCRIBE
socket in order to avoid missing incoming events. While this first
implementation worked as intended, it didn't take into account the use
case when a second session/graph is deployed over existing, already
connected node managers; in those cases the system would try to issue a
new subscription through the SUBSCRIBE socket but without activity
happening on the corresponding monitor socket (a connection is
never made because it already exists). This in turn means that the
threading Event used to alert the caller about the connecting being
ready is never set, leading in turn to an error raised on the caller
side alerting about the situation.

This commit adds support for keeping track of the endpoints the
SUBSCRIBE socket is already connected to. This information is used to
then simply skip re-subscription to these endpoints, quickly letting the
caller know that the subscription is established.

A new unit test (for which existing infrastructure had to be slightly
adjusted) showcases the scenario (starts two successive sessions on the
same set of node managers). This is representative of the original
problem, and will ensure it doesn't happen in the future.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>