Add workingDir parameter to DockerApp
Docker containers run by default under /. Additionally, when we ensure
and create a new user in the container, we run under /home/r ("r" is the
new user). Both of these options make it impossible to retrieve anything
from the working directory after execution finishes though, which is why
we need a new parameter that lets users specify under which directory
the container command will run (and thus they can also map it from the
host system).

This commit adds support for a new workingDir parameter in DockerApps.
Its value defaults to "/", and is only taken into account when the image
from which the container is started doesn't already contain a value for
its WorkingDir specification. We take care that this preference is
obeyed both when we internally create a new user, as well as when we
don't A new unit test checks that this is working as expected.

This work is part of LIU-53.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>