Fix documentation generation for daliuge
There were a couple of warnings during documentation generation due to
missing dependencies not stated in the list of mocked modules. There
were also a warning due to a missing directory that was stated in the
configuration, but not otherwise used.

In this commit I've added the missing dependencies to the list of mocked
modules, taking the opportunity to rearrange the list elements to
there's one per line. I've also removed the configuration bit that
referred to the missing directory.

After these changes the documentation generates without warnings.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Former-commit-id: 0e1bce82b5958e391373a6867fe6c6eae6d6f3c0
Former-commit-id: f5e5e469051fbcd0ff7a7f322be5ab7d11344f8a