Incorporate ray docs into installation document
The document describing the merging between ray and daliuge was dangling
and not referenced by the main documentation. This produced a warning
when generating the documentation.

This commit simply includes the "README ray.rst" document at the end of
the "installation.rst" document, and fixes some section formatting in
the ray document that was wrong.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>