Add initial support for Branch drops
Branch drops are a special type of application drops that must implement
an additional condition function, returning true or false, and thus
triggering one branch of execution down the rest of the graph, or the
other. The current implementation requires these applications
to have exactly two outputs, mainly because the information of which
branch should continue executing is transmitted through the output drop
onto its corresponding consumers instead of trying to contact the
consumers directly.

To support this new functionality we require a new SKIPPED state, which
automatically propagates through the graph. Unit tests check that this
propagation works through multiple levels, both for data and application
drops, such that their respective states change correctly (which is
important for the UI to display them correctly too).

The Web UI has also been updated to display skipped branches of
execution with a different color.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>