Support unwritten FileDROPs correctly
In FileDROP's setCompleted call we update the _size attribute by issuing
an os.stat call. This call can fail when the file is not present, which
in turn happens when none of its producers actually wrote anything into
it. Depending on the context, the error will be visible or not (e.g., it
might not appear in the logfiles if the call was issued on a thread with
no high-level try/catch+log block -- like those spawned by
AppDrop.execute).

This commit fixes this by catching errors from the os.stat call.
Additionally when the file has not been created we create it; this way
we prevent consuming applications to fail reading from a non-existing
file.

A unit test has been included to reproduce this setup in a simple way.

This problem was reported in LIU-111.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Former-commit-id: b704a8b331f09510627f47be57f62a6bcfd70ddc
Former-commit-id: 4ba857291e29bae881fcd4123d96fc00dc79bde4