Acquire GIL in dynlib_example2 as appropriate
We have been getting away with this code, but in reality we have always
been guilty of not taking the GIL in the "pythonic" dynlibs (i.e., those
offering a "init2" method rather than an "init" one). Until 3.8 no
errors have showed up, but in 3.9 segfaults start to happen when calling
PyLong_FromLong (and possibly other methods).

To solve this I've updated our example code to take the GIL as
appropriate. This is verbose, but it's the simples solution. A different
approach would have been to load the dynlib with ctypes's PyDLL rather
than CDLL, but that forces all C function calls to hold the GIL even if
not required, which is less performant. Nobody (except Kevin, if he
still does?) is using this anyway, and if so they would have run into
the same error as we did here, so there's no issue in clarifying that
the GIL must be taken by the dynlib code rather than by daliuge.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>