Augment kwargs with values from applicationArgs
When creating drops from a graph specification we used to use only the
top-level key-values to build the keyword map passed down to
applications, which then inspected it to obtain their user-provided
arguments. For some time now LGs (and their translation into PGs) have
separated how arguments are provided to applications though, grouping
them together inside an applicationArgs map instead of placing them at
the top-level of each drop specification. This means that applications
cannot access these values directly, and that the mechanism to declare
arguments at the class level (using dlg_string_param and friends)
doesn't pick them up.

This commit merges the information contained in applicationArgs into the
set of arguments passed down to drops via kwargs. This is enough for the
dlg_*_param declarations to pick these up.

This is part of YAN-968.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Former-commit-id: 2b28d4b64e67c630c1f595d08ea27c195b54503a
Former-commit-id: 0fe696177b2a564cba7c024c5ec3c66fb312b57c