Remove eager string formatting in log statements
Log message formatting should be left to the logging system instead of
being carried out by users; this is because in many cases, specially in
the lower logging levels like DEBUG, messages will actually be filtered
out and never make it to any of the sinks. This is sadly a very common
mistake one sees in the wild, and can cause real performance issues.

These might have been the slow logging reported in #142, but even if it
isn't the change is beneficial.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Former-commit-id: b95cab2896024984a6475716b1ddce1829541424
Former-commit-id: 9c6fde66c3a606c8df096a62735df3a6b5c9dae2