Raise correct exception coming from DlgProcess
When an exception is thrown in an application drop running in a
DlgProcess instance, it's correctly sent to the main daliuge node
manager process, but it was incorrectly handled: instead of raising the
exception object, the code was raising a 2-tuple with the exception and
the traceback. This 2-tuple object doesn't inherit from BaseException of
course, and thus the exception users received was always a TypeError due
to an invalid exception type being thrown.

This commit properly converts the 2-tuple into an exception object with
the correct traceback, which is then correctly raised, giving users the
true reason of the failure.

This bug was found while working on YAN-974.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>