Change error trigger for test_invalid
We used to open port 1 to trigger an error on the SocketListenerApp
because sockets under 1024 can only be opened by privileged users;
however, when running tests under root these assumptions are not true
anymore and the test hangs.

This commit changes the test so the application tries to open port
65536, which is out of range regardless of the user trying to use it.

This was found while trying out daliuge with python 3.10 inside a Docker
container, thus running things as root.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>