Redesign DLM constructor parameters and behavior
The old version of the DLM constructor was unnecessarily complicated,
reading parameters from **kwargs instead of receiving them by name. This
has been fixed, and the parameter names have been changed to snake_case
instead of the old pascalCase.

The two existing parameters that have changed are cleanupPeriod
(cleanup_period now) and checkPeriod (check_period now). The default
values of cleanup_period has also changed: instead of defaulting to 10 *
check_period, it now defaults to its own constant value, which happens
to be 10 times the default value of check_period; giving an explicit
check_period won't change the default value of cleanup_period though.

A change in behavior has also been introduced (although not noticeable
by default): if these periods are set to zero, the corresponding feature
is shut off. This way users can now opt out from automatic drop cleanup
even if the DLM is up and running.

Likewise, a new parameter has been introduced to control whether
replication of data drops is active or not. It defaults to True to match
the existing behavior, but users can opt out from this by setting it to
False.

This work is part of YAN-999.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>