Change default DLM construction values
The DLM, once created, has always had all its features turned on by
default. This can be overwhelming in many scenarios, specially when
replication of big amounts of data is involved.

This commit switches the DLM construction to turn off all features by
default, passing on the responsibility of turning them on to users. Unit
tests have been updated to turn the necessary features on only when
required (which was almost always the case, by accident). On the other
hand, the NodeManager currently offers a single flag to users to turn
the DLM on or off; we are thus turning all features on when users
requests the DLM. More fine-grained options for the NodeManager will be
introduced to toggle these switches individually.

This work is part of YAN-999.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>