Fix broken template variable expansion
In 126bfeb a lot of code was re-formatted, but the formatting tool
missed the fact that the dm.html file is not used as-is, but is rather a
template. In this template, variable expansion happens with double
curly braces, which must be together, but the automatic formatting
interpreted them as code block delimiters and placed them each on a new
line.

This bug was prevent the web UI from functioning correctly, with the
problem becoming clear once the browser developer console was fired up.

This was found while working on YAN-999.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>