Simplify parameter forwarding in NodeManager
The forwarding of parameters from NodeManager to NodeManagerBase relied
on the position of defaulted arguments of NodeManagerBase, and was
becoming difficult to maintain.

This commit removes all mentioning of forwarded parameters in
NodeManager, and simply forwards *args and **kwargs. This way, adding
new parameters to NodeManagerBase (where the real action happens)
doesn't mean we have to explicitly add them to NodeManager too.

This work is part of YAN-999.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>