Fire up the tasks of the DLM on the NodeManager
Until now, when previously used, the DLM was only used to produce
replicas of the data drops, as its two background tasks that
automatically expire and delete drop payloads were never actually
started. I looked in the git history to the best of my abilities, and
couldn't find any place where the startup() method of the DLM was
actually invoked other than the tests.

This commit fires up the DLM background tasks in the Node Manager, which
is done in the start() method like the rest of the background-starting
tasks.

This work is part of YAN-999.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>