Show DLM background task creation in logs
Without this hint in the logs it isn't obvious at a first glance that
the DLM is starting its background tasks, if any. Having this printed
out in the logs makes it clearer to the user that something is going on,
and at which rate.

This work is part of YAN-999.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>