Add fine-grained command line options to drive DLM
Previously a single --no-dlm command line option in the Node Manager
program specified whether the DLM would be started or not (historically
it was started by default, then both the command-line option and the
Node Manager parameter were turned off by default). Now that the DLM
accepts separate flags to control each of its features, this catch-all
command line is insufficient.

Three new command-line parameters have been added for each of the
features of the DLM. Their default values are such that data replication
is off by default, but automatic data removal is on by default. This
doesn't match either of the two previous default behaviors, but
something in the middle that is allegedly a good compromise: there is no
data explosion due to unexpected data replicas, and unnecessary data is
removed periodically.

The old --no-dlm command line argument has been left, but has been
marked as deprecated. Whenever it is given, the individual flags are set
such that the DLM is started without any of its features.

This work is part of YAN-999.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>